Add CoreEventMission
