+++ api/package.json
@@ -54,7 +54,7 @@
         "@simplewebauthn/server": "^13.3.0",
         "@sinclair/typebox": "^0.34.0",
         "@tak-ps/etl": "^10.0.0",
-        "@tak-ps/node-cot": "^14.49.0",
+        "@tak-ps/node-cot": "^14.50.0",
         "@tak-ps/node-safeurl": "^1.5.0",
         "@tak-ps/node-tak": "^12.23.0",
         "@turf/bbox": "^7.1.0",
